errores de btnsdelete y taskform resueltos

+++ src/firebase/firestore.js
@@ -87,13 +87,15 @@ window.addEventListener('DOMContentLoaded', async () => {
     });
 
     const btnsDelete = tasksContainer.querySelectorAll('.btn-delete');
-    btnsDelete.forEach((btn) => {
-      btn.addEventListener('click', ({ target: { dataset } }) => {
-        if (confirm('¿Estás segura de que deseas eliminar esta publicación?')) {
-          deleteTask(dataset.id);
-        }
+    if (btnsDelete) {
+      btnsDelete.forEach((btn) => {
+        btn.addEventListener('click', ({ target: { dataset } }) => {
+          if (confirm('¿Estás segura de que deseas eliminar esta publicación?')) {
+            deleteTask(dataset.id);
+          }
+        });
       });
-    });
+    }
 
     const btnsEdit = tasksContainer.querySelectorAll('.btn-edit');
     btnsEdit.forEach((btn) => {
@@ -113,24 +115,26 @@ window.addEventListener('DOMContentLoaded', async () => {
   });
 });
 
-taskForm.addEventListener('submit', (e) => {
-  e.preventDefault();
+if (taskForm) {
+  taskForm.addEventListener('submit', (e) => {
+    e.preventDefault();
 
-  const description = taskForm['task-description'];
+    const description = taskForm['task-description'];
 
-  if (description.value.trim() === '') {
-    alert('No se pueden publicar campos vacíos :(');
-  } else {
-    if (!editStatus) {
-      saveTask(description.value);
+    if (description.value.trim() === '') {
+      alert('No se pueden publicar campos vacíos :(');
     } else {
-      updateTask(id, {
-        description: description.value,
-      });
+      if (!editStatus) {
+        saveTask(description.value);
+      } else {
+        updateTask(id, {
+          description: description.value,
+        });
 
-      editStatus = false;
+        editStatus = false;
+      }
+
+      taskForm.reset();
     }
-
-    taskForm.reset();
-  }
-});
+  });
+}
